freeswitch: fix cloning the stable branch
This automatically clones from the remote branch instead of doing the branch switch
locally.

The approach introduced in https://github.com/plivo/plivoframework/pull/125 does not work. It was working in the environment I was working on because I automatically did somethings to test the changes convenient. I did *not* test for completely clean machines. The old way would fail, because it's a shallow clone. It only kept one version of the repo's history and *ONLY* for the master branch. This is why it would fail when it switched to the stable branch. I apologize for the trouble.